+++ phpmyfaq/faq.php
@@ -126,8 +126,6 @@ $faqVisits->logViews((int) $faqId);
 $question = $faq->getQuestion($faqId);
 if ($faqConfig->get('main.enableMarkdownEditor')) {
     $answer = $converter->convert($faq->faqRecord['content'])->getContent();
-} else {
-    $answer = $faqHelper->rewriteLanguageMarkupClass($faq->faqRecord['content']);
 }
 
 // Cleanup answer content first